:recycle: TileVisitor does not need to be a template anymore
